Move detailed configs to after/plugin
This is drawing from ThePrimeagen's 0 to LSP tut. The after/ setup
seemed gimmicky to me at first, but it's grown on me. Primarily, it
means we're not dependent on the plugin manager to configure all of our
plugins after installation, which will make it easier to migrate to
whatever comes after Lazy.nvim.

In addition for general config and such, dumping files in runtimepath
seems more in line with the vim way, as opposed to a monolithic codebase
all linked together with `require`s.

Schema for now (subject to revision):
- plugin list and basic settings go in init.lua
- general config that is complex enough to extract goes in after/plugin
- plugin-specific config goes in after/plugin/pconf